feat: Change developer (cls) to administrator (cls) & Move to administrator.py
