+++ src/login.py
@@ -3,6 +3,7 @@ from PIL import Image
 from random import randint
 from .signup import signup_interface
 from .dashboard import dashboard
+from .administrator import administrator
 
 icon__password = Image.open(rf"{DIR_PATH}\assets\icons\material icons\password.png")
 icon__security = Image.open(rf"{DIR_PATH}\assets\icons\material icons\security.png")
@@ -558,7 +559,7 @@ class login_interface:
 
                 # raise NotImplementedError('build a auth for developer')
                 self.window.withdraw()
-                developer_window = login_interface._developer_(self.window)
+                developer_window = administrator(self.window)
                 self.window.wait_window(developer_window.window__developer)
                 self.window.deiconify()
 
@@ -930,43 +931,3 @@ class login_interface:
                 auth__processing.place(x=166, y=442)
                 auth__processing.after(2000, auth__processing.place_forget)
                 auth__processing.after(2000, security_code_accepted)
-
-    class _developer_:
-
-        def __init__(self, parent: customtkinter.CTk):
-
-            self.window__developer = customtkinter.CTkToplevel(parent)
-            self.window__developer.geometry("1000x500")
-            self.window__developer.title("Developer")
-            self.window__developer.resizable(False, False)
-            self.frame__developer = customtkinter.CTkFrame(
-                self.window__developer, width=980, height=480
-            )
-            self.frame__developer.place(x=10, y=10)
-
-            def hover(_):
-                rame = customtkinter.CTkFrame(
-                    self.frame__developer, width=100, height=100
-                )
-                rame.place(x=300, y=100)
-
-            dev = customtkinter.CTkButton(
-                self.frame__developer,
-                text="",
-                image=customtkinter.CTkImage(
-                    light_image=icon__developer_mode_tv,
-                    dark_image=icon__developer_mode_tv,
-                    size=(10, 10),
-                ),
-                fg_color="transparent",
-                hover=False,
-                width=0,
-                height=0,
-            )
-            dev.place(x=89, y=100)
-            dev.bind("<Motion>", hover)
-            customtkinter.CTkButton(
-                self.frame__developer,
-                text="Exit",
-                command=self.window__developer.destroy,
-            ).place(x=200, y=300)
